+++ Changelog.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## master (unreleased)
+
+Nothing here yet.
+
 ## v11.0.1 (2020-09-11)
 
 - Add ISO code decorator to Catalonia calendar, thanks to @jbagot (#551).