+++ Changelog.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 - Improve coverage of Singapore calendar (#546).
+- Add optional school holidays to Netherlands calendar.
 
 ## v11.0.1 (2020-09-11)